# openapi/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Spring Application exposes Api to manage configuration for EC/PSP on the Nodo dei Pagamenti",
    "termsOfService": "https://www.pagopa.gov.it/",
    "title": "PagoPA API configuration",
    "version": "0.2.28"
  },
  "host": "127.0.0.1:8080",
  "basePath": "/apiconfig/api/v1",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/creditorinstitutions/{creditorinstitutioncode}/stations": {
      "get": {
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "description": "Organization fiscal code, the fiscal code of the Organization.",
            "in": "path",
            "maxLength": 50,
            "minLength": 0,
            "name": "creditorinstitutioncode",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CreditorInstitutionStationList"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          },
          "500": {
            "description": "Service unavailable",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "Authorization": []
          }
        ],
        "tags": [
          "Creditor Institutions"
        ],
        "operationId": "getCreditorInstitutionStations",
        "summary": "Get station details and relation info with creditor institution"
      },
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "description": "Organization fiscal code, the fiscal code of the Organization.",
            "in": "path",
            "maxLength": 50,
            "minLength": 0,
            "name": "creditorinstitutioncode",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreditorInstitutionStationEdit"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/CreditorInstitutionStationEdit"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          },
          "409": {
            "description": "Conflict",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          },
          "500": {
            "description": "Service unavailable",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "Authorization": []
          }
        ],
        "tags": [
          "Creditor Institutions"
        ],
        "operationId": "createCreditorInstitutionStation",
        "summary": "Create station details and relation info with creditor institution"
      }
    },
    "/creditorinstitutions/{creditorinstitutioncode}/stations/{stationcode}": {
      "delete": {
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "description": "Organization fiscal code, the fiscal code of the Organization.",
            "in": "path",
            "maxLength": 50,
            "minLength": 1,
            "name": "creditorinstitutioncode",
            "required": true,
            "type": "string"
          },
          {
            "description": "station code.",
            "in": "path",
            "maxLength": 50,
            "minLength": 0,
            "name": "stationcode",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          },
          "500": {
            "description": "Service unavailable",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "Authorization": []
          }
        ],
        "tags": [
          "Creditor Institutions"
        ],
        "operationId": "deleteCreditorInstitutionStation",
        "summary": "Delete a relation between creditor institution and station"
      },
      "put": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "description": "The fiscal code of the Organization to update",
            "in": "path",
            "maxLength": 50,
            "minLength": 1,
            "name": "creditorinstitutioncode",
            "required": true,
            "type": "string"
          },
          {
            "description": "station code.",
            "in": "path",
            "maxLength": 50,
            "minLength": 0,
            "name": "stationcode",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreditorInstitutionStationEdit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CreditorInstitutionStationEdit"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          },
          "500": {
            "description": "Service unavailable",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "Authorization": []
          }
        ],
        "tags": [
          "Creditor Institutions"
        ],
        "operationId": "updateCreditorInstitutionStation",
        "summary": "Update a relation between creditor institution and station"
      }
    },
...
    "/stations": {
      "get": {
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "default": 50,
            "description": "Number of elements on one page. Default = 50",
            "format": "int32",
            "in": "query",
            "name": "limit",
            "required": false,
            "type": "integer"
          },
          {
            "description": "Page number. Page value starts from 0",
            "format": "int32",
            "in": "query",
            "name": "page",
            "required": true,
            "type": "integer"
          },
          {
            "description": "Filter by broker",
            "in": "query",
            "name": "brokercode",
            "required": false,
            "type": "string"
          },
          {
            "description": "Filter by creditor institution",
            "in": "query",
            "name": "creditorinstitutioncode",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Stations"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          },
          "500": {
            "description": "Service unavailable",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "Authorization": []
          }
        ],
        "tags": [
          "Creditor Institutions"
        ],
        "operationId": "getStations",
        "summary": "Get paginated list of stations"
      },
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/StationDetails"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/StationDetails"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          },
          "409": {
            "description": "Conflict",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          },
          "500": {
            "description": "Service unavailable",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "Authorization": []
          }
        ],
        "tags": [
          "Creditor Institutions"
        ],
        "operationId": "createStation",
        "summary": "Create a station"
      }
    },
    "/stations/{stationcode}": {
      "delete": {
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "description": "station code",
            "in": "path",
            "maxLength": 50,
            "minLength": 0,
            "name": "stationcode",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          },
          "500": {
            "description": "Service unavailable",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "Authorization": []
          }
        ],
        "tags": [
          "Creditor Institutions"
        ],
        "operationId": "deleteStation",
        "summary": "Delete a station"
      },
      "get": {
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "description": "station code.",
            "in": "path",
            "maxLength": 50,
            "minLength": 0,
            "name": "stationcode",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/StationDetails"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          },
          "500": {
            "description": "Service unavailable",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "Authorization": []
          }
        ],
        "tags": [
          "Creditor Institutions"
        ],
        "operationId": "getStation",
        "summary": "Get station details"
      },
      "put": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "description": "station code",
            "in": "path",
            "maxLength": 50,
            "minLength": 0,
            "name": "stationcode",
            "required": true,
            "type": "string"
          },
          {
            "description": "The values to update of the station",
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/StationDetails"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/StationDetails"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          },
          "500": {
            "description": "Service unavailable",
            "schema": {
              "$ref": "#/definitions/ProblemJson"
            }
          }
        },
        "security": [
          {
            "ApiKey": []
          },
          {
            "Authorization": []
          }
        ],
        "tags": [
          "Creditor Institutions"
        ],
        "operationId": "updateStation",
        "summary": "Update a station"
      }
    }
  },
  "definitions": {
...
    "CreditorInstitutionStation": {
      "properties": {
        "application_code": {
          "format": "int64",
          "type": "integer"
        },
        "aux_digit": {
          "format": "int64",
          "type": "integer"
        },
        "broadcast": {
          "type": "boolean"
        },
        "enabled": {
          "default": true,
          "description": "station enabled",
          "type": "boolean"
        },
        "mod4": {
          "type": "boolean"
        },
        "segregation_code": {
          "format": "int64",
          "type": "integer"
        },
        "station_code": {
          "example": "1234567890100",
          "maxLength": 35,
          "minLength": 0,
          "type": "string"
        },
        "version": {
          "description": "number version",
          "format": "int64",
          "type": "integer"
        }
      },
      "required": [
        "enabled",
        "station_code",
        "version"
      ],
      "type": "object"
    },
    "CreditorInstitutionStationEdit": {
      "properties": {
        "application_code": {
          "format": "int64",
          "type": "integer"
        },
        "aux_digit": {
          "format": "int64",
          "type": "integer"
        },
        "broadcast": {
          "type": "boolean"
        },
        "mod4": {
          "type": "boolean"
        },
        "segregation_code": {
          "format": "int64",
          "type": "integer"
        },
        "station_code": {
          "example": "1234567890100",
          "maxLength": 35,
          "minLength": 0,
          "type": "string"
        }
      },
      "required": [
        "station_code"
      ],
      "type": "object"
    },
    "CreditorInstitutionStationList": {
      "properties": {
        "stations_list": {
          "items": {
            "$ref": "#/definitions/CreditorInstitutionStation"
          },
          "type": "array"
        }
      },
      "required": [
        "stations_list"
      ],
      "type": "object"
    },
...
    "PageInfo": {
      "properties": {
        "items_found": {
          "description": "Number of items found. (The last page may have fewer elements than required)",
          "format": "int32",
          "type": "integer"
        },
        "limit": {
          "description": "Required number of items per page",
          "format": "int32",
          "type": "integer"
        },
        "page": {
          "description": "Page number",
          "format": "int32",
          "type": "integer"
        },
        "total_pages": {
          "description": "Total number of pages",
          "format": "int32",
          "type": "integer"
        }
      },
      "required": [
        "items_found",
        "limit",
        "page",
        "total_pages"
      ],
      "type": "object"
    },
...
    "ProblemJson": {
      "properties": {
        "detail": {
          "description": "A human readable explanation specific to this occurrence of the problem.",
          "example": "There was an error processing the request",
          "type": "string"
        },
        "status": {
          "description": "The HTTP status code generated by the origin server for this occurrence of the problem.",
          "example": 200,
          "format": "int32",
          "maximum": 600,
          "minimum": 100,
          "type": "integer"
        },
        "title": {
          "description": "A short, summary of the problem type. Written in english and readable for engineers (usually not suited for non technical stakeholders and not localized); example: Service Unavailable",
          "type": "string"
        }
      },
      "type": "object"
    },
...
    "Station": {
      "properties": {
        "enabled": {
          "default": true,
          "description": "station enabled",
          "type": "boolean"
        },
        "station_code": {
          "example": "1234567890100",
          "maxLength": 35,
          "minLength": 0,
          "type": "string"
        },
        "version": {
          "description": "number version",
          "format": "int64",
          "type": "integer"
        }
      },
      "required": [
        "enabled",
        "station_code",
        "version"
      ],
      "type": "object"
    },
    "StationDetails": {
      "properties": {
        "broker_code": {
          "type": "string"
        },
        "enabled": {
          "default": true,
          "description": "station enabled",
          "type": "boolean"
        },
        "flag_online": {
          "type": "boolean"
        },
        "ip": {
          "type": "string"
        },
        "ip_4mod": {
          "type": "string"
        },
        "new_password": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "port": {
          "format": "int64",
          "type": "integer"
        },
        "port_4mod": {
          "format": "int64",
          "type": "integer"
        },
        "protocol": {
          "type": "string"
        },
        "protocol_4mod": {
          "type": "string"
        },
        "proxy_enabled": {
          "type": "boolean"
        },
        "proxy_host": {
          "type": "string"
        },
        "proxy_password": {
          "type": "string"
        },
        "proxy_port": {
          "format": "int64",
          "type": "integer"
        },
        "proxy_username": {
          "type": "string"
        },
        "redirect_ip": {
          "type": "string"
        },
        "redirect_path": {
          "type": "string"
        },
        "redirect_port": {
          "format": "int64",
          "type": "integer"
        },
        "redirect_protocol": {
          "type": "string"
        },
        "redirect_query_string": {
          "type": "string"
        },
        "service": {
          "type": "string"
        },
        "service_4mod": {
          "type": "string"
        },
        "station_code": {
          "example": "1234567890100",
          "maxLength": 35,
          "minLength": 0,
          "type": "string"
        },
        "thread_number": {
          "format": "int64",
          "type": "integer"
        },
        "timeout_a": {
          "format": "int64",
          "type": "integer"
        },
        "timeout_b": {
          "format": "int64",
          "type": "integer"
        },
        "timeout_c": {
          "format": "int64",
          "type": "integer"
        },
        "version": {
          "description": "number version",
          "format": "int64",
          "type": "integer"
        }
      },
      "required": [
        "broker_code",
        "enabled",
        "ip",
        "password",
        "port",
        "protocol",
        "service",
        "station_code",
        "thread_number",
        "timeout_a",
        "timeout_b",
        "timeout_c",
        "version"
      ],
      "type": "object"
    },
    "Stations": {
      "properties": {
        "page_info": {
          "$ref": "#/definitions/PageInfo"
        },
        "stations": {
          "items": {
            "$ref": "#/definitions/Station"
          },
          "type": "array"
        }
      },
      "required": [
        "page_info",
        "stations"
      ],
      "type": "object"
    },
...
  },
  "securityDefinitions": {
    "ApiKey": {
      "description": "The API key to access this function app.",
      "in": "header",
      "name": "Ocp-Apim-Subscription-Key",
      "type": "apiKey"
    },
    "Authorization": {
      "bearerFormat": "JWT",
      "description": "JWT token get after Azure Login",
      "scheme": "Bearer",
      "type": "apiKey",
      "in": "header",
      "name": "Authorization"
    }
  },
...
}